requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.8

faiss-cpu==1.8.0.post1

pypdf==5.1.0
unstructured==0.16.4
markdown==3.7

rank-bm25==0.2.2

python-dotenv==1.0.1
tqdm==4.66.6
requests==2.32.3

streamlit==1.41.1
pandas==2.2.3